+++ src/components/BarChart/hooks/tests/use-y-scale.test.tsx
@@ -12,6 +12,7 @@ jest.mock('d3-scale', () => ({
     scale.range = (range: any) => (range ? scale : range);
     scale.domain = (domain: any) => (domain ? scale : domain);
     scale.nice = () => scale;
+    scale.copy = () => scale;
     return scale;
   }),
 }));
@@ -29,6 +30,7 @@ describe('useYScale', () => {
       scale.range = (range: any) => (range ? scale : range);
       scale.domain = (domain: any) => (domain ? scale : domain);
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -60,6 +62,7 @@ describe('useYScale', () => {
       domainSpy = jest.fn((domain: any) => (domain ? scale : domain));
       scale.domain = domainSpy;
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -91,6 +94,7 @@ describe('useYScale', () => {
       domainSpy = jest.fn((domain: any) => (domain ? scale : domain));
       scale.domain = domainSpy;
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -122,6 +126,7 @@ describe('useYScale', () => {
       domainSpy = jest.fn((domain: any) => (domain ? scale : domain));
       scale.domain = domainSpy;
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -153,6 +158,7 @@ describe('useYScale', () => {
       scale.range = rangeSpy;
       scale.domain = (domain: any) => (domain ? scale : domain);
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -178,6 +184,7 @@ describe('useYScale', () => {
       scale.range = (range: any) => (range ? scale : range);
       scale.domain = (domain: any) => (domain ? scale : domain);
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });