+++ src/components/LineChart/hooks/tests/use-y-scale.test.tsx
@@ -13,6 +13,7 @@ jest.mock('d3-scale', () => ({
     scale.range = (range: any) => (range ? scale : range);
     scale.domain = (domain: any) => (domain ? scale : domain);
     scale.nice = () => scale;
+    scale.copy = () => scale;
     return scale;
   }),
 }));
@@ -30,6 +31,7 @@ describe('useYScale', () => {
       scale.range = (range: any) => (range ? scale : range);
       scale.domain = (domain: any) => (domain ? scale : domain);
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -58,6 +60,7 @@ describe('useYScale', () => {
       domainSpy = jest.fn((domain: any) => (domain ? scale : domain));
       scale.domain = domainSpy;
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -95,6 +98,7 @@ describe('useYScale', () => {
       domainSpy = jest.fn((domain: any) => (domain ? scale : domain));
       scale.domain = domainSpy;
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -131,6 +135,7 @@ describe('useYScale', () => {
       scale.range = rangeSpy;
       scale.domain = (domain: any) => (domain ? scale : domain);
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });
 
@@ -157,6 +162,7 @@ describe('useYScale', () => {
       scale.range = (range: any) => (range ? scale : range);
       scale.domain = (domain: any) => (domain ? scale : domain);
       scale.nice = () => scale;
+      scale.copy = () => scale;
       return scale;
     });