+++ src/utilities/index.ts
@@ -19,3 +19,4 @@ export {getPathLength} from './get-path-length';
 export {getPointAtLength} from './get-point-at-length';
 export {normalizeData} from './normalize-data';
 export {createCSSGradient} from './create-css-gradient';
+export {shouldRoundScaleUp} from './should-round-scale-up';